neo4j-local-setup/requirements.txt
neo4j>=5.28.0
python-dotenv>=1.0.1
fastapi>=0.115.0
uvicorn>=0.30.0

